Additional Classes and references fix
