+++ Dynamics/SimulationParameters.py
@@ -1,6 +1,12 @@
 import GromacsOutput
 import GromacsInput
-
+import pymol_plugin_dynamics
+import ProgressStatus
+import os
+try:
+    import prody
+except ModuleNotFoundError:
+    prody = False
 class SimulationParameters:
     gmx_output = ""
     gmx_input = ""
@@ -19,10 +25,10 @@ class SimulationParameters:
         if prody:
             self.vectors_prody = Vectors()
             print("ProDy correctly imported")
-        self.progress = ProgressStatus()
+        self.progress = ProgressStatus.ProgressStatus()
 
     def create_cfg_files(self):
-        self.em_file, self.pr_file, self.md_file = create_config_files(self.project_name)
+        self.em_file, self.pr_file, self.md_file =  pymol_plugin_dynamics.create_config_files(self.project_name)
 
     def change_stop_value(self, value):
         if value:
@@ -32,7 +38,7 @@ class SimulationParameters:
 
     def change_project_name(self, name):
         self.project_name = name
-        project_dir = get_project_dirs(self.project_name)
+        project_dir = pymol_plugin_dynamics.get_project_dirs(self.project_name)
         if not os.path.isdir(project_dir):
             os.makedirs(project_dir)
